add errors package
